Support filters directly in the config.js file
This requires to use a different method to stringify the plugin config
but also introduces some limitations: filters must be arrow functions
and can not use global variables or functions. However, it make easier
configuring this plugin and should still fit most use cases. If more
complex logic is needed, the separate file can be reintroduced later.

Note that it also impacts the `exclude` option which now supports a
regular expression object instead of a string pattern.